requirements.txt
streamlit==1.24.0
pandas==1.5.3
scikit-learn==1.3.0
joblib==1.3.1
Pillow
